// types.ts
export type AIProvider = 'gemini' | 'ollama';
export type StorageProvider = 'google-drive' | 'local-storage' | 'dropbox' | 'one-drive';

export interface Settings {
  aiProvider: AIProvider;
  ollamaModel: string;
  storageProvider: StorageProvider;
  storagePath: string;
}